pyproject.toml
[tool.poetry]
name = "tweeter"
version = "0.1.0"
description = ""
authors = ["Fred Melchior"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.0"
djangorestframework = "^3.14.0"
django = "^4.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
